// src/shared/types.ts
export enum WeaponType {
  AssaultRifle = 'AR',
  Rifle = 'RF',
  MarksmanRifle = 'MMR',
  SubmachineGun = 'SMG',
  LightMachineGun = 'LMG',
  Shotgun = 'SG',
  Pistol = 'Pistol',
}

export enum WeaponRarity {
  Exotic = 'exotic',
  Named = 'named',
  HighEnd = 'high-end',
}

export enum GearRarity {
  Exotic = 'exotic',
  Named = 'named',
  HighEnd = 'high-end',
  GearSet = 'gear-set',
}

export enum GearStatType {
  Offensive = 'offensive',
  Defensive = 'defensive',
  Skills = 'skills',
}

export enum GearSlot {
  Mask = 'mask',
  Backpack = 'backpack',
  BodyArmor = 'body_armor',
  Gloves = 'gloves',
  Holster = 'holster',
  KneePads = 'knee_pads',
}